main.cpp: take version information from autotool configuration
The autotool configuration needs to know the version of the project, so
instead of forcing the developer to maintain two version strings and keep them
synchronized, allow the code to take the version string from the autotool
configuration automatically. Hence, only one version needs to be maintained
which is shared between the autotool configuration and the code.

Signed-off-by: Trevor Woerner <twoerner@gmail.com>